requirements.txt
streamlit
pytesseract
spacy
pillow
